Added AUCTeX with latemk support and latex-preview-pane
AUCTeX is setup with multifile support and knows to use latexmk.

Also set up latex-preview-pane to display pdf preview in side pane.
Set to use latexmk and update preview when file is saved, cool feature
but the pdf viewer is incredibly slow, if I can speed this up it would
be a nice feature but for now it's better to use an external viewer.