Turned on diff visualization (undo-tree), turn on auto-cache (predictive)
In undo tree turned on diff visualisation to show changes while browsing
tree.

In predictive mode config enabled caching of dictionary updates to only
be added during idle time.
Also found a bug in which the main dict (not in the working dir) are no
longer used after a percent is used in math mode: $\%$